+++ wrappers/os/ubuntu/HAL_OS_linux.c
@@ -32,22 +32,22 @@
 #define PLATFORM_WAIT_INFINITE (~0)
 
 #ifdef DYNAMIC_REGISTER
-char _product_key[IOTX_PRODUCT_KEY_LEN + 1]       = "a1ZETBPbycq";
-char _product_secret[IOTX_PRODUCT_SECRET_LEN + 1] = "L68wCVXYUaNg1Ey9";
-char _device_name[IOTX_DEVICE_NAME_LEN + 1]       = "example1";
-char _device_secret[IOTX_DEVICE_SECRET_LEN + 1]   = "";
+    char _product_key[IOTX_PRODUCT_KEY_LEN + 1]       = "a1ZETBPbycq";
+    char _product_secret[IOTX_PRODUCT_SECRET_LEN + 1] = "L68wCVXYUaNg1Ey9";
+    char _device_name[IOTX_DEVICE_NAME_LEN + 1]       = "example1";
+    char _device_secret[IOTX_DEVICE_SECRET_LEN + 1]   = "";
 #else
-#ifdef DEVICE_MODEL_ENABLED
-char _product_key[IOTX_PRODUCT_KEY_LEN + 1]       = "a1RIsMLz2BJ";
-char _product_secret[IOTX_PRODUCT_SECRET_LEN + 1] = "fSAF0hle6xL0oRWd";
-char _device_name[IOTX_DEVICE_NAME_LEN + 1]       = "example1";
-char _device_secret[IOTX_DEVICE_SECRET_LEN + 1]   = "RDXf67itLqZCwdMCRrw0N5FHbv5D7jrE";
-#else
-char _product_key[IOTX_PRODUCT_KEY_LEN + 1]       = "a1MZxOdcBnO";
-char _product_secret[IOTX_PRODUCT_SECRET_LEN + 1] = "h4I4dneEFp7EImTv";
-char _device_name[IOTX_DEVICE_NAME_LEN + 1]       = "test_01";
-char _device_secret[IOTX_DEVICE_SECRET_LEN + 1]   = "t9GmMf2jb3LgWfXBaZD2r3aJrfVWBv56";
-#endif
+    #ifdef DEVICE_MODEL_ENABLED
+        char _product_key[IOTX_PRODUCT_KEY_LEN + 1]       = "a1RIsMLz2BJ";
+        char _product_secret[IOTX_PRODUCT_SECRET_LEN + 1] = "fSAF0hle6xL0oRWd";
+        char _device_name[IOTX_DEVICE_NAME_LEN + 1]       = "example1";
+        char _device_secret[IOTX_DEVICE_SECRET_LEN + 1]   = "RDXf67itLqZCwdMCRrw0N5FHbv5D7jrE";
+    #else
+        char _product_key[IOTX_PRODUCT_KEY_LEN + 1]       = "a1MZxOdcBnO";
+        char _product_secret[IOTX_PRODUCT_SECRET_LEN + 1] = "h4I4dneEFp7EImTv";
+        char _device_name[IOTX_DEVICE_NAME_LEN + 1]       = "test_01";
+        char _device_secret[IOTX_DEVICE_SECRET_LEN + 1]   = "t9GmMf2jb3LgWfXBaZD2r3aJrfVWBv56";
+    #endif
 #endif
 
 char _firmware_version[IOTX_FIRMWARE_VER_LEN] = "app-1.0.0-20180101.1000";
@@ -93,14 +93,14 @@ uint32_t HAL_Random(uint32_t region)
     FILE *handle;
     ssize_t ret = 0;
     uint32_t output = 0;
-    handle = fopen("/dev/urandom","r");
+    handle = fopen("/dev/urandom", "r");
     if (handle == NULL) {
         printf("open /dev/urandom failed\n");
         return 0;
     }
-    ret = fread(&output,sizeof(uint32_t),1,handle);
+    ret = fread(&output, sizeof(uint32_t), 1, handle);
     if (ret != 1) {
-        printf("fread error: %ld\n",ret);
+        printf("fread error: %d\n", (int)ret);
         fclose(handle);
         return 0;
     }
